Feature #3111. Use now the invitMe Silverpeas plugin the the user zoom to send an invitation to the user.
In the integration of the invitMe plugin, in order to keep the same behaviour with the
invitation sending feature in others pages (profil, whitepages, directory), the userZoom
Silverpeas plugin takes care of both the relationships of the current user and of the
invitations sent by the current user; the current user being the one behind the current
HTTP session.
In order to keep great interactivity, the relationships and the sent invitations are
reloaded each time the userZoom popups a tooltip. Then, when an invitation is sent or
accepted, the information is kept up to date and the invitMe plugin won't be invoked
for the user that has received or accepted the invitation.

In order to keep track the relationships of the current user, the UserProfileResource REST
service takes into account now a new term for the user identifier: me. Latest means the
user behind the current session of the HTTP request. With a such particular value, clients
don't need to know the identifier of the current user to ask some particular information
about him.

In order to keep track the invitations sent or received by the current user, a new REST
service is written: InvitationResource. For instance, this WEB resource represents an
invitation and provides two access points to them: thoses sent by the current user and
accessible at services/invitations/outbox, and thoses received by the current user and
accessible at services/invitations/inbox.