Feature #3111. In order to facilitate the writting of tests for REST resources, simplify some test preparation and code, mainly with the OrganizationController that became a true mock (managed with mockito) and wrapped by an OrganizationControllerMockWrapper instance managed by the IoC container and that plays the role of an OrganizationController instance for business logic. (It delegates the method invocation to the wrapped mock).
Transform MockablePersonalizationService to a true PersonalizationService mock wrapper. It initializes now the mock and delegates to it all the method invocations. It plays the role of a PersonalizationService instance for the business logic involved in the tests. The wrapped mock can be get to register some behaviour expected by the tests.