Feature #6071 - actions after merge from 5.15.x: - fixing compilation problems - applying some move operations due to renaming of packages of some API - moving subscribe integration tests to awaiting folder - adding the possibility to add simple additional parameters to a resource event in order to retrieve them from a JMS listener
Migration of unit tests to integration tests around subscription will be done in a second time.

Modifying maven dependencies in order to provide an implementation of the specified version of JSTL API (actually, the aim is to get the JSTL TLD definitions).
By this way, our IDEs are capable to provide useful auto completion and help around JSTL tags.