Feature #5710 : implementing an HTML5 drag & drop plugin for file & folder uploads. Adding Promise Polyfill (ES6) : "es6-promise.min.js" version 3.0.2 Wiring silverpeas-fileUpload plugin with the new file upload WEB Service. Wiring the ImportDragAndDrop servlet that is used by Kmelia application. Wiring the DragAndDrop servlet that is used by Attachment services (deleting VersionedDragAndDrop servlet).
Modifying Drag & Drop help messages for following applications:
- Gallery
- Kmelia
- Attachment

Improving the popup resize algorithm.

Modifying import mechanism in order to take into account:
- content language
- version tracking
- if documents must be imported every one in a same publication or one per publication

Adding a new tool in FileUtil in order to move easily files from sub folder of a folder to the root of the folder.

Upgrading PublicationAccessController in order to handle all the rules about the access rights and so to use it everywhere... As a consequence, the SimpleDocumentAccessController has been wired with it.