Fix a bug that was discovered with the deletion of a user in Silverpeas: the data of the user in the Silverpeas's social network was never cleaned up. Now, when a user is deleted, its relationships are also deleted and hence he's also removed from the IM roster of the buddies of his (and his own roster is also cleaned up)
Add a default logging config for the chat engine in Silverpeas.

Fix a bug in which the online status wasn't considered as a presence status like the others statuses.
Now, when a disconnected user switches its presence status to one other than offline status then he's being connected with the selected presence status.

At user disconnection from Silverpeas, the disconnection mechanism is now performed within a promise (SilverChat.stop() returns a promise that is resolved
when the chat client is effectively stopped).