Feature #8017 Refactor the export and the import of calendar events to and from an iCalendar representation of them by implementing the Exporter and the Importer interfaces of the Import/Export API.
The interfaces Exporter and Importer in the Import/Export API have been
modified to take into account the lambda, a new feature in Java 8.

Move the process of the import of iCal events into the Calendar Engine into a
dedicated class: iCalendarEventImportProcessor. This class uses the service of
a ICalendarImporter instance to consume an iCalendar source and it is dedicated
to report the consumed events and occurrences into the persistence context of
Silverpeas (new events, new occurrences, updated events, updated occurrences).

Currently, the implementation of the import of iCal events into Silverpeas
follows the rules below:
- the import is performed for a given calendar existing in Silvereas
- each new imported event or event occurrences is persisted into the given
  calendar
- an existing event is updated from the same imported event whose the
  modification is posterior to the last update of the existing event.
- an existing event occurrence is updated from the same imported event
  occurrence whose the modification is posterior to the last update of the
  existing event occurrence
- the attendees aren't taken in charge; so if the user adds some attendees to
  an imported event or occurrence, this information is kept when it is updated
  by a new import with the same events or occurrences.

Some limitations in this mechanism are known; they concern the already imported
events and occurrences:
- the update of an event doesn't apply to its existing occurrences
- the remove of any modifications to an event occurrence isn't taken into
  account (the already imported event occurrencew are ever persisted)
- the deletion of an event isn't taken into account