[Feature 8017] Add a scheduling job to synchronize all the of synchronized calendars in Silverpeas.
The job is scheduled from the CRON expression defined by the property
'calendar.synchronization.cron' in the properties file
properties/org/silverpeas/calendar/settings/calendar.properties.
The synchronization of the calendars is multi-threaded. By default, the number
of threads is based upon the number of available processors in the server
running Silverpeas. Nevertheless, it is possible to fix this number with the
property 'calendar.synchronization.processors' in the same properties file.

When synchronizing all the synchronized calendars in Silverpeas, the result of
each calendar synchronization is reported into a log in the category defined
by the constant ICalendarEventSynchronization#REPORT_NAMESPACE. By default, the
log level of this logging namespace is set at 'INFO' in the logging
configuration file
properties/org/silverpeas/util/logging/calendarSynchroLogging.properties. It
is then possible to change this value either directly in this file or through
the backoffice in Silverpeas.

In order the Silverpeas JPA-based persistence framework could be used by any
system tasks (like the scheduled calendars synchronization here) for which
there is no specific requesting user, a little and simple modification was
done: it is now possible to set a user (a system user for example) for a system
task by using the OperationContext instance mapped with the current running
thread. When performing effectively the persistence operation, the Silverpeas
JPA-based framework asks for the user that has requested the operation in order
to set some technical fields in the entity to persist (the creator and/or the
udpater of the entity). Now, the OperationContext instance mapped with the
current running thread will check if there is yet a user set, otherwise it will
call for the user of the user session behind the operation. In the case of the
calendars synchronization triggered by a system task (like the scheduling), for
each calendar, the user in the OperationContext instance mapped with the
current thread is set with the calendar's creator.