Fix some issues with the new MVC mechanism.
Add a new AngularJS directive to render and manage attachments of a
given resource.

Update attachment code at creation time to hide details about the way the
uploaded files are attached. Move some interfaces on the contribution to the
core-api project. Add the Translation interface for both the old i18Bean and
the new Contribution objects and rename the class Translation in core-library
to BeanTranslation