Up to now, the URI of the web entities was computed by using the JAX-RS class UriInfo. The base URI returned by UriInfo objects is made up of the URI scheme, the host, the port if it isn't a standard one and the root path to which all the root web resources are relative. When Silverpeas is listening at a secure port, the base URI returned by the UriInfo objects is made up of the port 443 though this port number is a standard one for the scheme HTTPS and hence it should be missed in the URI. When this occurs, our security mechanism block all such request URIs because they differ from the web domain of Silverpeas (in the web, the port declaration is a part of a domain); it is considered as a cross-domain attack.
In order to fix this behaviour, the use of UriInfo is now replaced by a new
class: WebResourceUri. This class wraps UriInfo to provide some of its
features (like the getting of the URI parameters, for example). This new class
takes also into account the ApplicationURL Silverpeas property and computes
correctly the web domain (scheme + host + optionally the port) by using the
method URLUtil#getFullApplicationUrl(HttpServletRequest). It provides also
additional methods to get the base URI or path of a running web resource (that
is a web resource that behaves for a given component instance if any).

A new class is also defined: SilverpeasWebResource. It represents a Web
resource in Silverpeas and it defines some static methods to get the base path
of all of the web resources defined in Silverpeas. WebResourceUri is also based
upon these methods. These methods can be also used to compute a more complete
URI for a given web entity.

With these changes a great majority of web resources in Silverpeas were updated
consequently.