Refactor the WYSIWYG: * The configuration is now performed through the WysiwygEditorConfig that is   available to the web component controllers by the abstract class   AbstractComponentSessionController. WysiwygEditorConfig provides a method   to format a configuration into a JSON representation expected by CKEditor. * The WYSIWYG editor configuration is now performed per application. So, it is   required to pass as argument in the WysiwygEditorConfig construction the name   of the concerned application. * The configuration of the WYSIWYG editor isn't more centralized into the   properties file org/silverpeas/wysiwyg/settings/wysiwygSettings.properties.   Each application can customize the configuration for itself by providing a   properties file in org/silverpeas/wysiwyg/settings whose the name starts by   the application name (for example for infoLetter:   infoLetterSettings.properties). In such a properties file they can override   some of the properties like the configFile and specify with the property   noCss if the Silverpeas stylesheets have to be or not applied to the editor   for the application instances. * The Wysiwyg class is now WysiwygEditor and it is used by WysiwygTag to   configure the editor for the current application instance (hence for the   application) and to render such configuration in JSON in the HTML pages. * The rich content that is authored with the WYSIWYG editor is now reified into   a WysiwygContent that refers the contribution to which it is related. The   WysiwygContent object can be saved or deleted by itself. For doing it uses   the service of a WysiwygContentRepository object that is managed by CDI.   The already existing WysiwygManager class implements now the interface   WysiwygContentRepository.
Remove the class ComponentHelper and add the methods directly into the
WAComponent class as static ones. Now, to get the application name of a
component instance, juste write:
    WAComponent.getByInstanceId(componentInstanceId).get().getName()
or:
    WAComponent.getByInstanceId(componentInstanceId)
        .orElseThrow(() -> new ComponentNotFoundException(msg))
        .getName()

Unit tests and integration tests were fixed by taking into account the change
above.

Add a registry of WYSIWYG editor configurations. With such a registry, each
application has the opportunity to register their own configuration for the
WYSIWYG editor; usually they set only the toolbar name. By default, if no
configuration was registered for a given application, a default configuration
is returned by the registry. With the registry it is now possible to fetch the
WYSIWYG editor configuration for a given application in transverse UI services.
The registry is accessible to JSTL with the tag getWysiwygEditorConfigRegistry
in the TLD silverpeas functions (silverFunctions.tld).

The calendar engine provides an implementation of the Initialization interface
to set a specific WYSIWYG editor configuration for all the calendar based
applications (like Almanach for example). This configuration is then fetched by
the silverpeas-calendar-event-form-main.jsp template of the
silverpeas-calendar-event-form-main angular directive.

Use the angular plugin ng.Ckeditor that wraps the use of the famous CKEditor
javascript component. This plugin is automatically included in web pages with
the wysiwyg plugin.