Externalize the persistence technical operations into a dedicated class instead of OperationContext. For doing, a new class is defined: PersistanceOperation. Currently this class is extended into two JPA-specific implementations: JpaPersistOperation that represents an entity persistence operation and JpaUpdateOperation that represents an entity update operation. These two concretes classes are used by SilverpeasJpaEntity to set the creation and the update technical data (user and date) before the serialization itself and to memorize the value of the update and/or creation properties that are set explicitly (id est manually by the client). The PersistenceOperation instances are managed by OperationContext; they can be get or instanciated for the current thread. Two qualifiers are also defined: @UpdateOperation and @PersistOperation to qualify the implementation of a PersistenceOperation.
The behaviour about the management of the creation and of the update technical
data that are set manually is changed in order to be able to serialized
entities coming from an import process. Now, the creation and update properties
are decorallated and each of them can be set explicitly.

Another behaviour has changed: the author of a change (creation or update) is
now get from the datasource before the serialization itself so if it doesn't
exist an exception is thrown.