Contributions can be of different types according to the Silverpeas components (aka applications) that manage them. Some core contributions are defined like publications, comments or events; contributions that can be extended to a more specific use. Others are just specific to a Silverpeas components. Nevertheless all of them can be handled either in Silverpeas Core or in some other Silverpeas components without knowing the concrete type of such a contribution. In order to be handled in a more generic way, each contribution can returns a model object that gives an abstract view, a conceptual view, of itself. For doing a new interface is added to represent this conceptual view: ContributionModel. A default implementation, DefaultContributionModel, is also defined so that it is returned by default by the Contribution interface itself. This default implementation accesses the represented contribution by reflection. Any concrete type that which customize this default implementation can override the method Contribution#getModel() to return its specific implementation of a model.
Beside this new interface, a set of filters is added to apply filtering on some
properties to execute operations associated with the matched predicates.
The Filter interface is used by the ContributionModel in order to perform some
specific operations according to the type of a contribution business property.

These new interfaces are now used by DurationReminder to get the reference date
time from which the triggering date time of the reminder will be computed.
This reference date comes from a business property of the contribution related
by the duration reminder and the name of this property is set when the
triggering rule is set to the duration reminder.