With the integration of the fix of the bug #8929, some integration tests didn't pass anymore.
First, the user/group right management on nodes has slightly changed.
Indeed, when the rights of a user or of a group has to be replaced by those of
another user or of another group, the replacement can be also or not applied on
the nodes managed by the component instances that are concerned by such an
operation. (This situation occurs only if access rights can be applied on the
nodes, like the topics in Kmelia instances.) If the rights have to be also
applied on the nodes, then, actually, the previous access rights on the nodes
are then removed and replaced, if any, by the access rights of the other user
or group. Otherwise, the access rights on the nodes aren't taken into account
and therefore the existing ones for the targeted user or group are kept. This
is a change with the behaviour preceding the bug fix as previously any existing
access rights on nodes were always removed whatever the access rights
replacement has to be also or not applied on the nodes in the concerned
component instances.

Second, the access rights change on the spaces that are moved has slightly
changed. Before the bug fix, when a space is moved into another space, any
previously inherited user/group rights were removed before applying those of
the parent space (if any). With the bug fix, only the users or the groups in
the existing access right (represented by a SpaceProfileInst) are removed. This
was done for performance reason but now empty access rights can exist; this
break the expectation of the integration tests about the move of spaces into
another space without any access rights or without a specific access rights.
To restore the expectation, any empty existing access rights are removed in the
case of a space move.