The Variable API is refactored in order to be compliant with our coding rules based upon DDD (Domain Driven Design). With the refactoring, the VariablesWebManager doesn't perform anymore any business-like or persistence-like operations that should be performed by the Variable and VariableScheduledValue instances themselves.
The Variable API uses now the org.silverpeas.core.date.Period class that is
based upon the java.time API instead of the now deprecated
org.silverpeas.core.date.period.Period class. For doing, new methods are added
in the TemporalConverter and the org.silverpeas.core.date.Period classes.

Add support of min and max date/datetime in the date/datetime attribute
converter in our Persistence API based upon JPA.

Support failure messaging in the VariablesWebManager. Take into account also of
the Not Found treatment when a variable or a value doesn't exist.

Rename VariablePeriod and VariablePeriodEntity into respectively
VariableScheduledValue and VariableScheduledValueEntity.

Adds an integration test that covers the simple management of the Variables and
their scheduled values. From the test, the design of the Variable API has been
improved.

Fixes a lot of issues coming from the first design of the API (previous commit)
that was a refactoring of the old dynamic values feature.