Fix some integration tests.
Now that Wildfly 13 is used as the default application server for integration tests,
Some stuffs have to be done in order to work correctly with this application server.
Among them, the JMS queues defined in Silverpeas (and not in tests) are now directly
defined into the application server dedicated to the integration tests. Another point
is to define some stub services only one time and share them by the different tests
that require them.
Thanks to Wildfly 13, we can improve our way to test.