Feature #9884 Now, the subject of a notification is by default defined by the property GML.st.notification.subject in the general localization bundle in Silverpeas. This property can be overriden in each Silverpeas component by setting in their own bundle the property custom.st.notification.subject. By default, this property name is returned by the method AbstractTemplateUserNotificationBuilder#getBundleSubjectKey().
The value of these two properties are a StringTemplate expression. In order to
access the properties of the resource about which a notification is sent, in
the subject setting, the object representing that resource is now passed into
the StringTemplate instance used by the notification builder.

To keep the flexibility in the setting of the notification subject with the
notification builder, the method getBundleSubjectKey() can be overriden to
specify a custom property other than custom.st.notification.subject.