Feature #9884 Improve by cleaning the HTML content of the messager popup.
Fix some bugs and remove the use of AlertUser in the session controller base
classes in prevision of the AlertUser removing.

SilverStatistics use now the messager to send a message to one connected user
or to all connected users.

Remove the javascript plugin messageme. It is now replaced by messager.
UserZoom uses now messager in place of messageme.

Fix some bugs