Bug #10857
In the user notification transverse module, the receivers that can be selected
can be now filtered also by the resource they can access and for which the
notification is sent. Such resource can be for example a node that is a generic
way to represent a category or a topic to categorize the contributions.

Replace Thread.sleep by Awaitility in MassiveAsynchronousNotificationIT