Bug #9194
ProfileInst defines a profile of right access on either a component instance or
on a resource managed by that component instance (for instance, only nodes).
In the last case, the resource covered by the right access profile is referred
by three attributes: one valued with the local identifier of the resource,
another one valued with the type of the resource and the last one giving the
component instance to which the resource belongs. The last attribute defines
also the component instance that is covered by the profile. The two first
attributes are now structured into a single object: ProfiledObjectId. Codes
using previously one or all of the two first properties of ProfileInst are
updated to take care of this new object.

When a right access profile of a component instance or of one of its node is
deleted or updated then a CDI notification is sent. The notification event is
then caught by a listener in the subscription transverse module. The goal of
the listener for instance is to unsubcribe any users that are subscribed to
a resource for which an access right profile was deleted; usually the deletion
of such a profile comes from the deletion of the resource.