Bug #9194
ProfileInst defines a profile of right access on either a component instance or
on a resource managed by that component instance (for instance, only nodes).
In the last case, the resource covered by the right access profile is referred
by three attributes: one valued with the local identifier of the resource,
another one valued with the type of the resource and the last one giving the
component instance to which the resource belongs. The last attribute defines
also the component instance that is covered by the profile. The two first
attributes are now structured into a single object: ProfiledObjectId. Codes
using previously one or all of the two first properties of ProfileInst are
updated to take care of this new object.

When a right access profile of a component instance or of one of its node is
updated then a CDI notification is sent. The notification event is then caught
by a listener in the subscription transverse module. The goal of the listener
is to unsubcribe any users that were removed from the updated profile. In the
case the user is removed from a given profile of a component instance, then
any subscriptions of this user on the resources managed by that component
instance are also deleted. This is applied also when a group is removed from a
profile: any users in that group and of its subgroups will be targeted by the
unsubscription process. The listener caught also notification events about
the deletion of a right access profile to apply automatically unsubscription on
the component instance or on the resource that is concerned by the deletion;
usually such a profile is deleted when the resource covered by this profile is
deleted.