Bug #9194
One bug was discovered while fixing bug #9194. Some applications like Kmelia
manages some type of resources that can be covered by some explicit right
access profiles in which the users and the groups able to access the component
instance are more specifically filtered. For example the folders (represented
by nodes) in Kmelia. In those applications, when a user or a group is removed
from a right access profile (ProfileInst) of a space or of an instance of such
an application, then the removing isn't propagated down to the resources (the
folders in Kmelia). The fix is about to ensure to propagate the removing when
it concernes a space or a component instance and not specifically a resource
managed by a component instance.