Bug #9194
The behaviour of the user subscription in Kmelia is changed to behave like any
others Silverpeas applications. So, the data in the 'subscribe' database table
is updated as consequently.

Fix an issue in the migration script orphanSubscriptionCleanup.
Add a migration script to remove any access rights on topics in Kmelias when
the user or the group behind those access rights have no more access the
Kmelia application.