Bug #10884
The way the shortcut to a publication is working is now more coherent.

Previously, if an additional location for a given publication is set, depending
on the component instance of that location, the new location was or not
considered in the GUI as a shortcut, although in the code any locations of a
publication is ... an alias (even if the location is the main one of the
publication).

In order to accomplish this coherence, the previously Alias class is renamed to
Location and as its name suggests a Location object refers the node in which
a publication can be located. In the case this location isn't the main one of
a publication, this location is named alias (for shortcut) of the main
location. This means that for a publication a location is either a main one
and it be changed only by path-based operations (copy, move, ...) or it is a
location and it can be changed only by relocation operations (unless its father
has been copied or moved).So, a publication is an alias to an original one if
its location isn't the main one.

The method in PublicationService to get now locations and aliases were
modified to take into account the father of a publication is a location that
can be either the main one or an alias. By this new meaning, the calls of those
methods were also rewritten because it is more now explicit to get the aliases
of a publication and because de locations carry now the information of being
or not an alias by themselves.