Bug #9194
Refine the way the control of the access of a user or a group to a given
resource is done. The control uses now the different AccessController objects.
For doing, the control of the access of a group is now in the AccessControllers
used by the notification mechanism.

In the subscription subsystem, a control is performed on the access of the
subscriber to the subscribed resource. Indeed, in some circumstances, when the
access right of a user or a group on a resource is removed, the subscription
of that user or group on the resource can be not removed. Because of the way
the access rights are managed currently in the Silverpeas GUI, it is not
possible to fix this behaviour without adding any strong complexity in the
code. This is why the decision was to check a subscription is valid or not.
The subscription validation is done by verifying the access rights of the
subscriber on the subscribed resource.

To implement the control of access of a group to a resource in Silverpeas
(a component instance or an object in a component instance), the API of the
Admin was lightly modified (Administration, OrganizationController,
AdminController, ...). For example, the method isComponentAvailable was renamed
to isComponentAvailableToUser in order to add a new method,
isComponentAvailableToGroup. In some circumstances, the original method was
kept but marked as deprecated.