Bug #10993
Split the setting of the SQL parameters in a PreparedStatement into several
setter objects, each of them dedicated to set a specific parameter type.
This is to avoid the use of multiple if-conditions and it allow to add more
easily an additional setter without having to modify the implementation of the
JdbcSqlExecutor interface.

Improve the setter of SQL date by taking explicitly into account the type of the
parameter value is of type java.sql.Date.