Bug #11135
Now, by default, the search of users by criteria doesn't exclude the removed
ones. A removed user is a user that is a deletion awaiting; its account isn't
then deleted and it can be restored later with the right access profiles.
It is then now required to explicitly exclude the removed users in the search.
By doing so, no more errors are thrown when we are looking for users with among
them some whose account in Silverpeas is removed; this is the case with the
user subscriptions for example.

However, the web service accessing the users maintains its original behaviour
that is to look only for users that are neither deleted nor removed because
its main use is to send back users with which a current user can interact with.
But, in some circumstances, the users that are also removed can be asked. For
such a context, a new query parameter is provided: includeRemoved.