Bug #11139
The configuration properties of the Silverpeas Chat client are now wrapped in
the ChatSettings class.
Some properties in the chat.properties configuration file are renamed whereas
another ones are added:
- chat.servers.xmpp to indicate the base URL of the XMPP server
- chat.xmpp.httpUpload to indicate the base URL of the HTTP file transfer
  service of the XMPP server when this one is listening at another URL than the
  one indicated by chat.servers.xmpp property.
A new class is introduced: ChatInitialization. It's goal is to set the security
settings of Silverpeas with the URL(s) of the services published by the XMPP
server. For doing, the SecuritySettings provides now a Registration object
whose aim is to allow any external components in Silverpeas to register some
URLs as to be security valid.