Completing silverpeas.js with new method whenSilverpeasEntirelyLoaded(callback).
Now, whenSilverpeasReady and whenSilverpeasEntirelyLoaded returns both a promise, so the given callback is optional.
The given callback if any is invoked when the promise is resolved.

The promise of whenSilverpeasReady is invoked when the DOM have been parsed and loaded but not yet all sub-ressources (images, CSS, etc.).

The promise of whenSilverpeasEntirelyLoaded is invoked when the DOM and all sub-ressources (images, CSS, etc.) have been parsed and loaded.