Feature #Virtual-Keyboard
Use Simple-Keyboard Javascript lib to integrate in Silverpeas a virtual
keyboard. This lib is wrapped by our own Javascript plugin, SilverKeyboard.
For instance, SilverKeyboard is enabled only if the following property in the
org/silverpeas/general.properties 'web.tool.virtualKeyboard' is set to true.
Once enabled, the virtual keyboard is displayed and active only for textual
inputs and textarea HTML elements. However, due to some limitations and for
the original purpose of the virtual keyboard, it doesn't work for instance
for WYSIWYG edition.