Feature #11841
Does first a basic CMIS implementation for Silverpeas using the OpenCMIS
libraries in order to have a glance of the way to implement CMIS and of the
potential difficulties we could encounter.
Fix the issue that broke the access the Silverpeas CMIS repository through the
Web Service (SOAP) bridge.

A CMIS object model is defined in which each type is a mapping of of a resource
in Silverpeas having to be exposed in the CMIS objects tree. The model has as
main class CmisObject from which all others types have to be extended. The object
model defines the properties expected by CMIS for each type defined in the model.
Requires some additional stuff to have a better Silverpeas resource <-> CMIS
object mapping.

The CMIS objects tree is rooted at a virtual collaborative space with as id "WA0".
The children of this virtual space are all the root collaborative spaces in
Silverpeas. Each object type in our CMIS object model has to specify the type of
the children it accepts (as expected by CMIS).
Now all the collaborative spaces and all the component instances in Silverpeas
can be accessed through the CMIS API.
For doing, start to create an abstract and generic way to walk across the
virtual CMIS object tree in Silverpeas by using different walkers, one for
each type of Silverpeas resources. For instance: a walker for spaces and
another one for application instances. The walker structure is based upon
an adaptation of both the Visitor, Factory and Chain of Responsability patterns.

Introduce a new interface Identifiable that should be implemented by the
business classes used by our CMIS support implementation. By this way we
can refine our implementation to be a little more generic when handling
such business objects. SpaceInst, SpaceInstLight, ComponentInst and
ComponentInstLight satisfy this new interface.

Put in place an authentication mechanism in our CMIS implementation that uses
the Silverpeas's authentication engine for REST web services.
Put in place an authorization mechanism that uses our Access Controller
framework.

Add unit tests to cover the functionalities of our CMIS support
implementation. For doing, a virtual Silverpeas environment is created for
those tests.

Move the CMIS model into core-library.
Filter the component instances that are exposed through CMIS to those of
Kmelia.