Bug #11365: making complex string templates customizable.
When a String is produced with StringTemplate API from a set of template files (not a single file), the customization was not effective when just a few (not all) of template files were overridden.
The detection of the template file directory (the one of the product or the one of customizations) was not rightly done because the customization needs all template files customized and not just a few.

A modification has been done in order to permits to customize just the needed template files. Now, when a custom folder exists and if it has been requested to StringTemplate API, the template files from product and the customized are merged into a new folder and the template files taken into account from string producing come from this new folder.

The new folder is called "__merged_delete_me_on_template_modification" and is located into the folder of customized template files.
When a customized template is modified, the new folder (if it already exists) MUST be deleted in order that the modifications are taken into account.