[quickinfo-optims] Increasing performances around the statistic services.
To get the number of access by contribution for a batch of contributions, services of statistics can now query the database by a single query for all contributions instead of performing a query per contribution.

Adding also a new parameter ticker.display.limit (from look properties) which permits to limit the number of news raquested by the ticker.