Bug #bug-11426
Resolve the issue by taking into account only the user domains
in Silverpeas that have a mapping to an XMPP domain defined in the
chat.properties settings file. The rendering of silverchat and the
connection opening with the remote XMPP server is only performed now if the
chat service is enabled for the current user (that is if the both the chat
service is enabled and the domain the user belongs to has a mapping to an XMPP
domain).

A default XMPP domain mapping can be now explictly set with the new property in
the chat.properties settings: chat.xmpp.domain.default. If this property is
set, any non explicitly mapped user domains in Silverpeas will be mapped to this
default XMPP domain. If not set, then the users in the non-mapped domains in
Silverpeas won't have the chat enabled for them.

Add unit tests on the user account registration in the chat server.